cpp warnings fix (CLion auto fix ugl) + rotation fix + mem cleanup

+++ Cpp/shader.cpp
@@ -10,7 +10,7 @@ std::string get_file_contents(const char* filename)
         in.seekg(0,std::ios::end);
         contents.resize(in.tellg());
         in.seekg(0,std::ios::beg);
-        in.read(&contents[0],contents.size());
+        in.read(&contents[0],static_cast<long long>(contents.size()));
         in.close();
         return(contents);
     }
